backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis[hiredis]>=5.0.0
python-multipart>=0.0.9
httpx>=0.27.0
# Sprint 4: Celery + Beat 스케줄러
celery[redis]>=5.3.0
# Sprint 9: Playwright 크롤러
playwright>=1.44.0
fake-useragent>=1.5.1
python-dotenv>=1.0.1
